corrigindo bugs do cadastro de post

+++ src/main/java/com/dminer/controllers/PostController.java
@@ -1,11 +1,13 @@
 package com.dminer.controllers;
 
+import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
+import java.util.regex.Pattern;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -24,15 +26,21 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.dminer.constantes.Constantes;
+import com.dminer.converters.CommentConverter;
+import com.dminer.dto.CommentDTO;
 import com.dminer.dto.PostDTO;
 import com.dminer.dto.PostRequestDTO;
+import com.dminer.entities.Comment;
 import com.dminer.entities.FileInfo;
 import com.dminer.entities.Post;
+import com.dminer.enums.PostType;
 import com.dminer.response.Response;
+import com.dminer.services.CommentService;
 import com.dminer.services.FileDatabaseService;
 import com.dminer.services.FileStorageService;
 import com.dminer.services.PostService;
 import com.dminer.services.UserService;
+import com.dminer.utils.UtilNumbers;
 
 import lombok.RequiredArgsConstructor;
 
@@ -58,14 +66,15 @@ public class PostController {
 	@Autowired
 	private UserService userService;
 
-	//@Autowired
-	//private CommentService commentService;
+	@Autowired
+	private CommentService commentService;
 	
-
+	@Autowired
+	private CommentConverter commentConverter;
     
-	@PostMapping(consumes = {"multipart/form-data"})
-	public ResponseEntity<Response<PostDTO>> create(
-		@RequestParam(value = "files", required = false) MultipartFile[] files,  @RequestBody PostRequestDTO postRequestDTO) {
+
+	@PostMapping(consumes = {"multipart/form-data", "application/json"})
+	public ResponseEntity<Response<PostDTO>> create( @RequestParam(value = "files", required = false) MultipartFile[] files,  @RequestBody PostRequestDTO postRequestDTO ) {
 		
 		log.info("----------------------------------------");
 		log.info("Salvando um novo post {}", postRequestDTO);
@@ -78,34 +87,53 @@ public class PostController {
 			return ResponseEntity.status(HttpStatus.NOT_FOUND).body(response);
 		}
 		
-		Post post = postService.persist(new Post());
+
+		Post post = new Post();
 		if (postRequestDTO.getContent() != null) 
 			post.setContent(postRequestDTO.getContent());
+		
+		if (UtilNumbers.isNumeric(postRequestDTO.getLikes() + ""))
+			post.setLikes(postRequestDTO.getLikes());
+
+		if (postRequestDTO.getType() != null && !postRequestDTO.getType().isEmpty())
+			post.setType(PostType.valueOf(postRequestDTO.getType()));
+		
+		List<Comment> comments = new ArrayList<>();
+		if (!postRequestDTO.getComments().isEmpty()) {
+			log.info("Adicionando comentários");
+			postRequestDTO.getComments().forEach(commentReq -> {
+				Comment comment = commentConverter.requestDtoToEntity(commentReq);
+				comment = commentService.persist(comment);
+				comments.add(comment);
+			});
+		}
+
+		post = postService.persist(post);
 
 		String caminhoAbsoluto;
 		try {
 			caminhoAbsoluto = criarDiretorio(post.getId());
-		} catch(RuntimeException e) {
-			postService.delete(post.getId());
+		} catch(IOException e) {
+			rollback(post, null, comments);
 			response.getErrors().add(e.getMessage());
 			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
 		}
 
 		List<FileInfo> anexos = new ArrayList<>();
 
-    	try {
-			Path path = Paths.get(caminhoAbsoluto);
+		Path path = Paths.get(caminhoAbsoluto);
+		if (files != null && files.length > 0) {
 			List<MultipartFile> array = Arrays.asList(files);			
-
+			
 			Optional<FileInfo> info = Optional.empty();
-
+			
 			for (MultipartFile multipartFile : array) {
 				String arquivoUrl = caminhoAbsoluto + "\\" + multipartFile.getOriginalFilename();
-
+				
 				if (fileStorageService.existsDirectory(Paths.get(arquivoUrl))) {
 					log.info("Diretório já existe no storage = {}", arquivoUrl);
 					info = fileDatabaseService.persist(new FileInfo(arquivoUrl, post));
-
+					
 				} else {
 					if(fileStorageService.save(multipartFile, path)) {
 						log.info("Salvando novo arquivo no storage {}", arquivoUrl);
@@ -117,18 +145,13 @@ public class PostController {
 					anexos.add(info.get());
 				}
 			}
-    	    
-			Post temp = postService.persist(post);
-			log.info("Adicionando anexos ao post e atualizando. {}", temp);
-			
-			response.setData(toDto(temp, anexos));
-			return ResponseEntity.status(HttpStatus.OK).body(response);
-
-		} catch (Exception e) {
-			rollback(post, anexos);
-			response.getErrors().add("Erro ao salvar post. Erro: " + e);
-			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(response);
-    	}
+		}
+		
+		Post temp = postService.persist(post);
+		log.info("Adicionando anexos ao post e atualizando. {}", temp);
+		
+		response.setData(postToDto(temp, anexos, comments));
+		return ResponseEntity.status(HttpStatus.OK).body(response);
 	}
     
 	
@@ -184,39 +207,62 @@ public class PostController {
 		}
 
 		Optional<List<FileInfo>> anexos = fileDatabaseService.findByPost(post.get());
-		
-		response.setData(toDto(post.get(), anexos.get()));
+		Optional<List<Comment>> comments = commentService.findByPost(post.get());
+		response.setData(postToDto(post.get(), anexos.get(), comments.get()));
 		return ResponseEntity.status(HttpStatus.OK).body(response);
 	}
 
 
-	private void rollback(Post post, List<FileInfo> anexos) {
-		anexos.forEach(anexo -> {
-			fileDatabaseService.delete(anexo.getId());
-		});
-		postService.delete(post.getId());
-		fileStorageService.delete(Paths.get(Constantes.appendInRoot(post.getId() + "")));
+	private void rollback(Post post, List<FileInfo> anexos, List<Comment> comments) {
+		if (anexos != null) {
+			log.info("Apagando os anexos do banco");		
+			anexos.forEach(anexo -> {
+				fileDatabaseService.delete(anexo.getId());
+			});
+		}
+
+		if (comments != null) {
+			log.info("Apagando os comentários do banco");
+			comments.forEach(comment -> {
+				commentService.delete(comment.getId());
+			});
+		}
+
+		if (post != null) {
+			log.info("Apagando o post do banco");
+			postService.delete(post.getId());
+			fileStorageService.delete(Paths.get(Constantes.appendInRoot(post.getId() + "")));
+		}
 	}
 
-
-	private PostDTO toDto(Post post, List<FileInfo> anexos) {
+	private PostDTO postToDto(Post post, List<FileInfo> anexos, List<Comment> comments) {
 		PostDTO dto = new PostDTO();
-		dto.setId(post.getId());
+		dto.setIdUsuario(post.getId());
+		dto.setLikes(post.getLikes());
+		dto.setType(post.getType().name());
+		dto.setId(post.getId());		
 		dto.setContent(post.getContent());
 		anexos.forEach(e -> {
 			dto.getAnexos().add(e.getUrl());
 		});
+
+		comments.forEach(e -> {
+			dto.getComments().add(commentConverter.entityToDTO(e));
+		});
 		return dto;
 	}
-	
+
 
 	/**
 	 * Cria diretórios organizados pelo id do Post
 	 */
-	private String criarDiretorio(int idPost) throws RuntimeException {
+	private String criarDiretorio(int idPost) throws IOException {
 		String diretorio = Constantes.appendInRoot(idPost + "");
-		log.info("Verificando se o diretório existe {}", diretorio);
+		log.info("Criando diretório {}", diretorio);
 		fileStorageService.createDirectory(Paths.get(diretorio));
 		return diretorio;
 	}
+
+	
+
 }
